add snapshot import test
